Add math functions, Enzyme compatibility, extended tests, benchmarks, and docs
Core changes:
- src/ctps.jl: allocating math functions (exp, log, sqrt, sin, cos, tan, asin,
  acos, sinh, cosh) with pool-backed in-place variants; fix exp/exp! early-
  return guard (degree_mask==0 instead of a0==zero(T)); remove desc field from
  CTPS struct, expose via virtual getproperty
- src/mathfunc.jl: pre-allocate term_next before loop for Enzyme compatibility
  (heap-only allocations, no stack-allocated temporaries)
- src/polymap.jl: polynomial map utilities
- src/macro.jl: new macro helpers
- src/TPSA.jl: updated exports

Tests (501/501 passing):
- test/arithmetic_tests.jl: 171 new accuracy tests (constructors, element,
  arithmetic operators, in-place ops, order truncation, dense multiplication)
- test/mathfunc_tests.jl: 258 tests for all math functions
- test/multiplication_tests.jl: strengthened with exact coefficient checks
- test/runtests.jl: include new test files; threadsafe and type_stability updated

Benchmarks:
- benchmarks/benchmark_GTPSA.jl: comparison vs GTPSA.jl
- benchmarks/benchmark_autodiff.jl: Enzyme differentiation benchmarks
- benchmarks/benchmark_mathfunctions.jl: math function timing

Docs and examples:
- docs/index.md, docs/api.md: Enzyme/AD interoperability section
- docs/tutorial.md: new tutorial
- README.md, docs/README.md, examples/README.md: updated
- examples/07_enzyme_ad.jl: 5 verified Enzyme+TPSA examples

+++ src/polymap.jl
@@ -13,7 +13,7 @@
 struct PolyMap
     dim::Int
     max_order::Int
-    map::Matrix{Int}
+    map::Matrix{UInt8}  # UInt8 sufficient: max exponent ≤ max_order < 256
 
     function PolyMap(dim::Int, order::Int)
         new(dim, order, setindexmap(dim, order))
@@ -78,11 +78,10 @@ end
 
 function setindexmap(dim::Int, max_order::Int)
     totallength = binomial(max_order + dim, dim)
-    # map = [decomposite(i, dim) for i in 0:totallength-1]
-    # create map as a matrix
-    map = zeros(Int, totallength, dim + 1)
+    # UInt8 is 8× more cache-friendly than Int64; max exponent ≤ max_order ≤ 127 in practice
+    map = zeros(UInt8, totallength, dim + 1)
     for i in 0:totallength-1
-        map[i + 1, :] = decomposite(i, dim)
+        map[i + 1, :] = decomposite(i, dim)  # Julia converts Int → UInt8 on assignment
     end
     return map
 end
